Created Google Integration Framework
Initial stage of the Google Integration Framework.

GoogleController will be the class that holds the required information
to connect to the specific Google API (Calendar or Tasks).

GoogleController will also maintain CalendarController and
TasksController.

CalendarController will connect to and interface with the Google
Calendar API.

Tasks Controller will connet to and interface with the Google Tasks API.

CalendarViewer will be providing static methods to format the Google
Calendar output.

TasksViewer will be providing static methods to format the Google Tasks
output.

GoogleAuthorizer provides a static method to setup and return a
authorized Credential object. It sets the scope of the Google API as
well as read in the developer’s Client’s information.