Update to TasksController's Logic
TasksController has been updated to use a specific TaskList ID so that
it will not create or work on duplicated Task Lists.

In this update, the task ID has been hardcoded into the TasksController
class to ensure that things are working. Therefore, it is still
incomplete as it requires a further step to save the TaskList ID to the
host’s computer.

If the TasksController is unable to grab any valid TaskList from the
Google Account, it will simply create one and store the newly created
TaskList’s ID into the host. (to be implemented)