Fixed a boundary case; Created TasksHandler
First, a boundary case related to getting the “ChirpTaskv0.1” TaskList
has been fixed. This problem arises when the ID of the previously
created TaskList has been corrupted or tampered with for example.

Second, a TasksHandler class has been created which is mainly used to
store static methods related to Google Tasks. Usage include checking if
a TaskList is null for now. (example: TasksHandler.isNull(TaskList)).