Implement Update; Overload Add Task; [GoogleTasks]
Implemented more methods in TasksHandler to support more Google Tasks’s
Task properties such as “set completed” (mark as done).

Implemented the updateTask method in TasksHandler which does the actual
execution of updating the task to Google Tasks.

Overloaded the addTask method in TasksController to take in a variation
of task. Previously only supported to take in a “floatingTask” title
value. Now includes a permutation of “floatingTask”, “notes”, “due
date”.