Added methods to show undone&hidden [Google Tasks]
Hidden Tasks are completed tasks and are cleared off the list. By
default, the normal tasks list shows undone and completed tasks until
the clear has been issued and executed.

Undone tasks are not completed, yet to be marked as done, in the
current tasks list. This method allows for filtering such as
filter:undone