Added creation of ChirpTask Calendar
This commit allows ChirpTask to create and maintain a Google Calendar
by its Calendar ID if there isn’t one that was created by ChirpTask
before.

If there is already a ChirpTask calendar, it would be used as the main
calendar to work on.

By making the above possible, CalendarHandler class is created to
provide static methods for CalendarController to call to make actions
related to Google Calendar service.

ChirpTask will store the Calendar ID in its newly created file
“credentials/googlecalendar/ChirpTaskTimedTaskCalendarID.txt”