Setup asynchronous framework for Google Integration
Google calls will now be executed concurrently so that it will not
affect the JavaFX GUI responsiveness.

Two new classes are introduced: ConcurrentController and
ConcurrentHandler.

ConcurrentController is in-charge of the ExecutorService, Callable and
Future, handling the separate thread all together.

ConcurrentHandler is in-charge of maintaining the callable classes for
the Google operations such as add (ConcurrentAdd).

Added new method: addGoogleIdToStorage(Task) in GoogleStorage.
This method is added to adapt to the concurrency changes.

Added String googleId to chirptask.storage.Task ADT, along with the
respective getter and setter methods.
This allows the methods to set the Google ID to the Task in the future.
It is currently used in ConcurrentAdd.