Update filter keyword function
Previous commit treat the enter filter string as the keyword string to
be searched. This should not be the case. If it does not fulfil any of
the “/<filter>”, the individual string will now be used as a keyword to
filter.

This is repeated until the end of the entire “filters” string.